Extends watson instead of overriding everything
