+++ qwatson/mainwindow.py
@@ -22,7 +22,7 @@ from PyQt5.QtWidgets import (QApplication, QGridLayout, QHBoxLayout,
 
 # ---- Local imports
 
-from qwatson.watson.watson import Watson
+from qwatson.watson.watsonextends import Watson
 from qwatson.utils import icons
 from qwatson.utils.watsonhelpers import round_frame_at
 from qwatson.widgets.clock import ElapsedTimeLCDNumber
