--- qwatson/watson/__init__.py
@@ -1,4 +1,0 @@
-from .watson import __version__  # noqa
-from .watson import Watson, WatsonError
-
-__all__ = ['Watson', 'WatsonError']
