--- qwatson/watson/version.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-version = "1.6.0"
